added coins added the number of coins in the game added the logic "when 100 coins, add 1 life" added the number of lives (starting with 3 lives) added the initial view and the start button added the life consumption and the death screen added the inputs

+++ Assets/_OprRunner/Scripts/Environment/Obstacles/ObstaclesSpawner.cs
@@ -9,7 +9,7 @@ namespace _OprRunner.Scripts.Environment.Obstacles
         [SerializeField] private GameObject[] _obstaclesPrefab;
         [SerializeField] private Transform _spawnPoint;
         [SerializeField] private Transform _obstaclesHolder;
-        [SerializeField] private int _spawnTimer;
+        [SerializeField] private float _spawnTimer;
 
         private Coroutine _spawnRoutine;
